+++ packages/expo-router/src/react-navigation/material-top-tabs/__tests__/missing-dependency.test.ios.tsx
@@ -9,6 +9,7 @@ jest.mock(
 );
 
 test('throws an error when react-native-tab-view is not installed', () => {
+  jest.resetModules();
   jest.doMock('react-native-tab-view', () => {
     throw new Error();
   });